pyproject.toml
[project]
name = "chatbot-embedding-tools"
version = "0.1.0"
description = "A Flask-based chat interface"
requires-python = ">=3.12,<4.0"

dependencies = [
    "flask>=3.0.0",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.10,<0.4.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.hatch.metadata]
allow-direct-references = true
